Refactors proxy logic into dedicated module
Moves the proxy request handling logic into a separate module for better organization and separation of concerns.

This improves code maintainability and allows for easier testing of the proxy functionality.
